More cleaning up of code, added shared info file, and seperate radial_menu file
